dildovod_app/requirements.txt
streamlit==1.28.0
pandas==2.0.3
python-docx==0.8.11
PyPDF2==3.0.1
python-multipart==0.0.6
jinja2==3.1.2
openpyxl==3.1.2